Dao tests

+++ utopia/src/main/java/com/ss/training/utopia/dao/AirportDAO.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface AirportDAO extends JpaRepository<Airport, Long> {
-	Airport findByAirportId(Integer airportId);
+	Airport findByAirportId(Long airportId);
 }
